Mudança de int para double nos tipos de produção das cidades Implementação da produção de recursos, tendo em conta a produtividade da célula, o número de trabalhadores e a quantidade de produção Agora os tipos de terreno tem produções diferentes, montanhas tem mais ouro mas menos comida, e planicie o contrario

+++ src/main/java/com/mycompany/projeto2/poo/ITerrain.java
@@ -15,7 +15,7 @@ public interface ITerrain {
     public int getEntryCost();
     public int getCyclesToTraverse();
     public int getMaxNumWorkers();
-    public int getFoodProduction();
-    public int getIndustrialProduction();
+    public double getFoodProduction();
+    public double getIndustrialProduction();
     public double getGoldProduction();
 }
